Tests/LibWeb/Text/expected/query-scroll-size-of-inline-node.txt

Scroll Height: 0px, Scroll Width: 0px
